Bugfix #166
