+++ gulpfile.mjs
@@ -11,8 +11,6 @@ import uglify from 'gulp-uglify';
 import bump from 'gulp-bump';
 import semver from 'semver';
 import gulpFile from 'gulp-file';
-import i18next from 'gulp-i18next-conv';
-import through2 from 'through2';
 import { exec } from 'child_process';
 
 const { src, dest, series, watch } = gulp;
@@ -176,47 +174,6 @@ export function makepot() {
 }
 
 
-// Funktion zur Anpassung der JSON-Struktur an das WordPress-Format
-function formatForWordPress() {
-    return through2.obj(function (file, _, cb) {
-        if (file.isBuffer()) {
-            let jsonContent = JSON.parse(file.contents.toString());
-
-            // WordPress-kompatible Struktur
-            let wpJson = {
-                locale_data: {
-                    messages: {
-                        "": {
-                            "domain": "rrze-faudir"
-                        }
-                    }
-                }
-            };
-
-            // JSON-Schlüssel hinzufügen
-            Object.keys(jsonContent).forEach(key => {
-                wpJson.locale_data.messages[key] = [jsonContent[key]];
-            });
-
-            // Aktualisierte JSON-Struktur speichern
-            file.contents = Buffer.from(JSON.stringify(wpJson, null, 2));
-        }
-        cb(null, file);
-    });
-}
-
-
-// Funktion zur Konvertierung von .po nach .json
-export function convertPoToJson() {
-    return src("languages/*.po")
-        .pipe(i18next())
-	.pipe(formatForWordPress())
-        .pipe(dest('languages/'))
-	.pipe(touch());
-};
-
-
-
 // Watch-Task zur Überwachung von Änderungen
 // Der Watch-Task wird zur Entwicklung verwendet, daher hier nur 
 // Dev-Versionen erstellen
@@ -294,8 +251,6 @@ export function createReadme() {
         });
 }
 
-
-export const pot = series(makepot);
 
 // Nur den Block neu bauen
 export const buildBlock = series(buildFaudirBlock);
