+++ package.json
@@ -46,7 +46,6 @@
         "@wordpress/scripts": "^30.12.0",
         "gulp": "^5.0.0",
         "gulp-concat": "^2.6.1",
-        "gulp-i18next-conv": "^12.0.0",
         "gulp-sass": "^5.0.0",
         "gulp-wp-pot": "^2.5.0",
         "sass": "^1.77.8"
@@ -60,7 +59,6 @@
         "gulp-sourcemaps": "^2.6.5",
         "gulp-touch-cmd": "^0.0.1",
         "gulp-uglify": "^3.0.2",
-        "semver": "^7.6.3",
-        "through2": "^4.0.2"
+        "semver": "^7.6.3"
     }
 }
